requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
Flask-Bcrypt==1.0.1
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
Flask-Talisman>=1.1.0
Flask-Limiter>=3.5.0
PyMySQL==1.1.0
python-dotenv==1.0.0
APScheduler==3.10.4
pandas>=2.0.3
openpyxl>=3.1.2
joblib>=1.3.0
scikit-learn>=1.3.0
numpy>=1.24.0
